server_2/requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.21.1
cmake==4.0.3
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
croniter==6.0.0
cryptography==45.0.6
csvw==3.5.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
deepface==0.0.79
dlinfo==2.0.0
docopt==0.6.2
ecdsa==0.19.1
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
encodec==0.1.1
espeakng==1.0.3
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
g2pkk==0.1.2
gast==0.4.0
gdown==5.2.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
grpcio==1.73.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
hangul-romanize==0.1.0
httptools==0.6.4
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
jax==0.4.30
jaxlib==0.4.30
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.1
jose==1.0.0
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
llvmlite==0.43.0
lxml==6.0.0
lz4==4.4.4
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.14
minio==7.2.16
ml_dtypes==0.5.1
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.6.4
murmurhash==1.0.13
namex==0.1.0
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.22.0
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==1.5.3
pdfminer.six==20250506
pdfplumber==0.11.7
phonemizer==3.3.0
Pillow==9.2.0
piper-tts==1.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
proglog==0.1.12
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
py-cpuinfo==9.0.0
py-espeak-ng==0.1.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyjokes==0.8.3
PyJWT==2.9.0
pynndescent==0.5.13
pyparsing==3.2.3
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pypinyin==0.55.0
pyreadline3==3.5.4
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2022.5
PyYAML==6.0.2
rdflib==7.1.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.11.4
seaborn==0.13.2
segments==2.3.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart_open==7.3.0.post1
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.7
soxr==0.5.0.post1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.1
starlette==0.46.2
SudachiDict-core==20250515
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-cpu==2.9.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.22.1
tqdm==4.67.1
trainer==0.0.36
transformers==4.55.3
typeguard==4.4.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2023.3
tzlocal==5.3.1
ultralytics==8.3.170
ultralytics-thop==2.0.14
umap-learn==0.5.7
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
uv==0.6.4
uvicorn==0.35.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.14.2
yarl==1.20.1
